git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/InductiveVisualMiner/Trunk@21865 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/plugins/inductiveVisualMiner/performance/QueueLengths.java
@@ -10,5 +10,5 @@ public interface QueueLengths {
 	 * @param time, use new Date(time) to make a Date-object
 	 * @return the number of cases in queue for this unode
 	 */
-	public long getQueueLength(UnfoldedNode unode, long time);
+	public double getQueueLength(UnfoldedNode unode, long time);
 }
